rename namespace to descriptor
