+++ model/src/main/java/de/fraunhofer/iosb/ilt/faaast/service/model/descriptor/IdentificationDescriptor.java
@@ -12,30 +12,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.fraunhofer.iosb.ilt.faaast.service.model.registry;
+package de.fraunhofer.iosb.ilt.faaast.service.model.descriptor;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import io.adminshell.aas.v3.model.Key;
-import io.adminshell.aas.v3.model.Reference;
+import io.adminshell.aas.v3.model.Identifier;
 import io.adminshell.aas.v3.model.builder.ExtendableBuilder;
 import java.io.Serializable;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Objects;
 
 
 /**
- * Registry Descriptor for Reference.
+ * Registry Descriptor for Identification.
  */
-public class ReferenceDescriptor implements Serializable {
+public class IdentificationDescriptor implements Serializable {
 
-    @JsonIgnore
     private String id;
-    private List<KeyDescriptor> keys;
+    private String idType;
 
-    public ReferenceDescriptor() {
+    public IdentificationDescriptor() {
         id = null;
-        keys = new ArrayList<>();
+        idType = null;
     }
 
 
@@ -49,13 +44,13 @@ public class ReferenceDescriptor implements Serializable {
     }
 
 
-    public List<KeyDescriptor> getKeys() {
-        return keys;
+    public String getIdType() {
+        return idType;
     }
 
 
-    public void setKeys(List<KeyDescriptor> keys) {
-        this.keys = keys;
+    public void setIdType(String idType) {
+        this.idType = idType;
     }
 
 
@@ -67,15 +62,15 @@ public class ReferenceDescriptor implements Serializable {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        ReferenceDescriptor that = (ReferenceDescriptor) o;
+        IdentificationDescriptor that = (IdentificationDescriptor) o;
         return Objects.equals(id, that.id)
-                && Objects.equals(keys, that.keys);
+                && Objects.equals(idType, that.idType);
     }
 
 
     @Override
     public int hashCode() {
-        return Objects.hash(id, keys);
+        return Objects.hash(id, idType);
     }
 
 
@@ -83,7 +78,7 @@ public class ReferenceDescriptor implements Serializable {
         return new Builder();
     }
 
-    public abstract static class AbstractBuilder<T extends ReferenceDescriptor, B extends AbstractBuilder<T, B>> extends ExtendableBuilder<T, B> {
+    public abstract static class AbstractBuilder<T extends IdentificationDescriptor, B extends AbstractBuilder<T, B>> extends ExtendableBuilder<T, B> {
 
         public B id(String value) {
             getBuildingInstance().setId(value);
@@ -91,29 +86,22 @@ public class ReferenceDescriptor implements Serializable {
         }
 
 
-        public B keys(List<KeyDescriptor> value) {
-            getBuildingInstance().setKeys(value);
+        public B idType(String value) {
+            getBuildingInstance().setIdType(value);
             return getSelf();
         }
 
 
-        public B key(KeyDescriptor value) {
-            getBuildingInstance().getKeys().add(value);
-            return getSelf();
-        }
-
-
-        public B from(Reference reference) {
-            if (reference != null) {
-                for (Key key: reference.getKeys()) {
-                    getBuildingInstance().getKeys().add(KeyDescriptor.builder().from(key).build());
-                }
+        public B from(Identifier identifier) {
+            if (identifier != null) {
+                getBuildingInstance().setId(identifier.getIdentifier());
+                getBuildingInstance().setIdType(identifier.getIdType().name());
             }
             return getSelf();
         }
     }
 
-    public static class Builder extends AbstractBuilder<ReferenceDescriptor, Builder> {
+    public static class Builder extends AbstractBuilder<IdentificationDescriptor, Builder> {
 
         @Override
         protected Builder getSelf() {
@@ -122,8 +110,8 @@ public class ReferenceDescriptor implements Serializable {
 
 
         @Override
-        protected ReferenceDescriptor newBuildingInstance() {
-            return new ReferenceDescriptor();
+        protected IdentificationDescriptor newBuildingInstance() {
+            return new IdentificationDescriptor();
         }
     }
 }
